Integrate content moderation components into a functional plugin
This commit integrates the previously implemented Azure Content Safety
moderator with the plugin architecture to moderate post content.

With this change, when posts are created or updated, they are added to
a queue of posts to be processed by the moderator. If a post fails
moderation, it is deleted.

In a future change, a bot will notify users of this deletion.