aOnActivate should succeed when not configured
Previously, if the plugin was not configured, OnActivate would return
an error. This would result in an error loop on the server where the
plugin manager would continously attempt to restart the plugin.

With this change, OnActivate will always return nil when the plugin is
not fully configured, or if moderation is explicitly disabled.

If moderation _is_ enabled and the plugin is not fully configured, we
still return nil; however, we will also log an error. This will help
in the event that the configuration is inadvertently changed, and some
required field is omitted.