Drop agent plugin validation on plugin start
We don't have any guarantees that the Agent plugin will be running
before this on server start - and in that case our initialization will
fail and we won't moderate any content. We don't need to do this
validation, as if there are problems, we'll see errors in the logs
when we try to analyze content.